+++ apps/web-client/src/app/interview/waiting/[sessionId]/components/InfoItem.tsx
